// src/app/interfaces/interfaces.ts
// Generated by https://quicktype.io

export interface RespuestaPost {
    ok: boolean;
    pagina: number;
    posts: Post[];
}

export interface Post {
    imgs?: string[];
    _id?: string;
    mensaje?: string;
    coords?: string;
    usuario?: Usuario;
    created?: string;
}

export interface Usuario {
    avatar?: string;
    _id?: string;
    nombre?: string;
    email?: string;
    password?: string;
}